# Support dynamic class constant fetch