--- a/composer.json
+++ b/composer.json
@@ -7,7 +7,7 @@
   "keywords": ["module", "xp"],
   "require" : {
     "xp-framework/core": "^11.0 | ^10.0",
-    "xp-framework/ast": "dev-feature/expressions as 10.0.0",
+    "xp-framework/ast": "dev-feature/expressions as 9.99.0",
     "php" : ">=7.0.0"
   },
   "require-dev" : {