--- a/src/main/php/lang/ast/emit/NullsafeAsTernaries.class.php
+++ b/src/main/php/lang/ast/emit/NullsafeAsTernaries.class.php
@@ -12,13 +12,6 @@ trait NullsafeAsTernaries {
     $result->out->write('null===('.$t.'=');
     $this->emitOne($result, $instance->expression);
     $result->out->write(')?null:'.$t.'->');
-
-    if ('literal' === $instance->member->kind) {
-      $result->out->write($instance->member->expression);
-    } else {
-      $result->out->write('{');
-      $this->emitOne($result, $instance->member);
-      $result->out->write('}');
-    }
+    $this->emitOne($result, $instance->member);
   }
 }