--- a/src/main/php/lang/ast/emit/PHP70.class.php
+++ b/src/main/php/lang/ast/emit/PHP70.class.php
@@ -1,7 +1,7 @@
 <?php namespace lang\ast\emit;
 
 use lang\ast\Node;
-use lang\ast\nodes\{InstanceExpression, ScopeExpression, Literal, Variable};
+use lang\ast\nodes\{Expression, InstanceExpression, ScopeExpression, Literal, Variable};
 use lang\ast\types\{IsUnion, IsIntersection, IsFunction, IsArray, IsMap, IsNullable, IsValue, IsLiteral, IsGeneric};
 
 /**
@@ -74,6 +74,9 @@ class PHP70 extends PHP {
       $this->emitOne($result, $callable->expression->expression);
       if ($callable->expression->member instanceof Literal) {
         $result->out->write(',"'.trim($callable->expression->member, '"\'').'"');
+      } else if ($callable->expression->member instanceof Expression) {
+        $result->out->write(',');
+        $this->emitOne($result, $callable->expression->member->inline);
       } else {
         $result->out->write(',');
         $this->emitOne($result, $callable->expression->member);
@@ -90,6 +93,9 @@ class PHP70 extends PHP {
       }
       if ($callable->expression->member instanceof Literal) {
         $result->out->write(',"'.trim($callable->expression->member, '"\'').'"');
+      } else if ($callable->expression->member instanceof Expression) {
+        $result->out->write(',');
+        $this->emitOne($result, $callable->expression->member->inline);
       } else {
         $result->out->write(',');
         $this->emitOne($result, $callable->expression->member);