--- a/src/main/php/lang/ast/emit/PHP83.class.php
+++ b/src/main/php/lang/ast/emit/PHP83.class.php
@@ -1,0 +1,45 @@
+<?php namespace lang\ast\emit;
+
+use lang\ast\Node;
+use lang\ast\types\{IsUnion, IsIntersection, IsFunction, IsArray, IsMap, IsNullable, IsValue, IsLiteral, IsGeneric};
+
+/**
+ * PHP 8.3 syntax
+ *
+ * @see  https://wiki.php.net/rfc#php_83
+ */
+class PHP83 extends PHP {
+  use RewriteBlockLambdaExpressions, ReadonlyClasses;
+
+  /** Sets up type => literal mappings */
+  public function __construct() {
+    $this->literals= [
+      IsArray::class        => function($t) { return 'array'; },
+      IsMap::class          => function($t) { return 'array'; },
+      IsFunction::class     => function($t) { return 'callable'; },
+      IsValue::class        => function($t) { return $t->literal(); },
+      IsNullable::class     => function($t) {
+        if (null === ($l= $this->literal($t->element))) return null;
+        return $t->element instanceof IsUnion ? $l.'|null' : '?'.$l;
+      },
+      IsIntersection::class => function($t) {
+        $i= '';
+        foreach ($t->components as $component) {
+          if (null === ($l= $this->literal($component))) return null;
+          $i.= '&'.$l;
+        }
+        return substr($i, 1);
+      },
+      IsUnion::class        => function($t) {
+        $u= '';
+        foreach ($t->components as $component) {
+          if (null === ($l= $this->literal($component))) return null;
+          $u.= '|'.$l;
+        }
+        return substr($u, 1);
+      },
+      IsLiteral::class      => function($t) { return $t->literal(); },
+      IsGeneric::class      => function($t) { return null; }
+    ];
+  }
+}