--- a/src/main/php/lang/ast/emit/RewriteClassOnObjects.class.php
+++ b/src/main/php/lang/ast/emit/RewriteClassOnObjects.class.php
@@ -1,6 +1,6 @@
 <?php namespace lang\ast\emit;
 
-use lang\ast\nodes\Literal;
+use lang\ast\nodes\{Expression, Literal};
 
 /**
  * Rewrites `[expr]::class` to `get_class($object)` except if expression
@@ -9,6 +9,7 @@ use lang\ast\nodes\Literal;
  * @see  https://wiki.php.net/rfc/class_name_literal_on_object
  */
 trait RewriteClassOnObjects {
+  use RewriteDynamicClassConstants { emitScope as rewriteDynamicClassConstants; }
 
   protected function emitScope($result, $scope) {
     if ($scope->member instanceof Literal && 'class' === $scope->member->expression && !is_string($scope->type)) {
@@ -16,7 +17,7 @@ trait RewriteClassOnObjects {
       $this->emitOne($result, $scope->type);
       $result->out->write(')');
     } else {
-      parent::emitScope($result, $scope);
+      $this->rewriteDynamicClassConstants($result, $scope);
     }
   }
 }