--- a/src/test/php/lang/ast/unittest/emit/InstantiationTest.class.php
+++ b/src/test/php/lang/ast/unittest/emit/InstantiationTest.class.php
@@ -27,6 +27,30 @@ class InstantiationTest extends EmittingTest {
   }
 
   #[Test]
+  public function new_dynamic_var() {
+    $r= $this->run('class <T> {
+      public function run() {
+        $class= \\util\\Date::class;
+        $var= "class";
+        return new $$var();
+      }
+    }');
+    Assert::instance(Date::class, $r);
+  }
+
+  #[Test]
+  public function new_var_expr() {
+    $r= $this->run('class <T> {
+      public function run() {
+        $class= \\util\\Date::class;
+        $var= "class";
+        return new ${$var}();
+      }
+    }');
+    Assert::instance(Date::class, $r);
+  }
+
+  #[Test]
   public function new_expr() {
     $r= $this->run('class <T> {
       private function factory() { return \\util\\Date::class; }