See https://wiki.php.net/rfc/dynamic_class_constant_fetch and https://github.com/php/php-src/pull/9793

* [x] Instance properties: `$this->$member`, `$this->{$member}`
* [x] Static properties: `self::${$member}`
* [x] Instance methods: `$this->$method()`, `$this->{$method}()`
* [x] Static methods: `self::$method()`, `self::{$method}()`
* [x] Constants: `self::{$member}`

This pull request also fixes support for dynamic static properties.