@thekid 2023-04-08
Performance and memory figures:

## Before
```
Test cases:  917 succeeded, 4 skipped
Memory used: 11181.34 kB (11235.46 kB peak)
Time taken:  0.262 seconds (0.760 seconds overall)
```

## After
```
Test cases:  933 succeeded, 4 skipped
Memory used: 11413.18 kB (11467.31 kB peak)
Time taken:  0.251 seconds (0.720 seconds overall)
```