@thekid 2023-04-08
Added to https://github.com/xp-framework/compiler/wiki/Implementation-status