@thekid 2023-04-08
Released in https://github.com/xp-framework/compiler/releases/tag/v8.10.0