@thekid 2023-04-08
The typical upgrade process will show:

  - Upgrading xp-framework/ast (v9.2.7 => v10.0.0): Extracting archive
  - Upgrading xp-framework/compiler (v8.9.0 => v8.10.0): Extracting archive
  - Upgrading xp-framework/reflection (v2.9.1 => v2.10.0): Extracting archive